[ALPS07524293] fget: check that the fd still exists after getting a ref to it
commit 054aa8d439b9185d4f5eb9a90282d1ce74772969 upstream.

Jann Horn points out that there is another possible race wrt Unix domain
socket garbage collection, somewhat reminiscent of the one fixed in
commit cbcf01128d0a ("af_unix: fix garbage collect vs MSG_PEEK").

See the extended comment about the garbage collection requirements added
to unix_peek_fds() by that commit for details.

The race comes from how we can locklessly look up a file descriptor just
as it is in the process of being closed, and with the right artificial
timing (Jann added a few strategic mdelay(500) calls to do that), the
Unix domain socket garbage collector could see the reference count
decrement of the close() happen before fget() took its reference to the
file and the file was attached onto a new file descriptor.

This is all (intentionally) correct on the struct file * side, with
RCU lookups and lockless reference counting very much part of the
design.  Getting that reference count out of order isnt a problem per
se.

But the garbage collector can get confused by seeing this situation of
having seen a file not having any remaining external references and then
seeing it being attached to an fd.

In commit cbcf01128d0a ("af_unix: fix garbage collect vs MSG_PEEK") the
fix was to serialize the file descriptor install with the garbage
collector by taking and releasing the unix_gc_lock.

Thats not really an option here, but since this all happens when we are
in the process of looking up a file descriptor, we can instead simply
just re-check that the file hasnt been closed in the meantime, and just
re-do the lookup if we raced with a concurrent close() of the same file
descriptor.

MTK-Commit-Id: 742cd489bde6b57c4a70b06509076a389247bd89

fget: check that the fd still exists after getting a ref to it
Reported-and-tested-by: Jann Horn <jannh@google.com>
Acked-by: Miklos Szeredi <mszeredi@redhat.com>
Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Signed-off-by: Qilin Tan <qilin.tan@mediatek.com>
CR-Id: ALPS07524293
Feature: F2FS
Change-Id: I730cdaa35a513041255e243e6a3ee9ac3483e8f8
(cherry-pick from 09f21027727f59807b15ce950d619ed934f87c60)
(cherry-pick from da8ecde33156f3ece1287211f8c94598bbfaee11)